+++ testdata/lang/functions/dst/main.c
@@ -16,7 +16,7 @@ static so_int sum3(so_int a, so_int b, so_int c) {
 int main(void) {
     so_int s0 = sum3(1, 2, 3);
     (void)s0;
-    main_Sum3Fn fn1 = sum3;
+    so_int (*fn1)(so_int, so_int, so_int) = sum3;
     so_int s1 = fn1(4, 5, 6);
     (void)s1;
     main_Sum3Fn fn2 = sum3;